Public endpoint to get tags added, safe deleting tag checking enabled

+++ src/main/kotlin/com/sphirye/lonjas/service/TagService.kt
@@ -3,11 +3,11 @@ package com.sphirye.lonjas.service
 import com.sphirye.lonjas.config.exception.DuplicatedException
 import com.sphirye.lonjas.config.exception.NotFoundException
 import com.sphirye.lonjas.entity.Tag
+import com.sphirye.lonjas.repository.PostRepository
 import com.sphirye.lonjas.repository.TagRepository
 import com.sphirye.lonjas.repository.criteria.TagCriteria
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.domain.Page
-import org.springframework.data.domain.PageImpl
 import org.springframework.stereotype.Service
 
 @Service
@@ -15,6 +15,7 @@ class TagService {
 
     @Autowired lateinit var tagRepository: TagRepository
     @Autowired lateinit var tagCriteria: TagCriteria
+    @Autowired lateinit var postRepository: PostRepository
 
     fun init() {
         if (tagRepository.count() <= 0) {
@@ -26,6 +27,7 @@ class TagService {
         if (existsByName(name)) { throw DuplicatedException("Theres already a tag with the given name.") }
         val tag = Tag()
         tag.name = name
+        tag.enabled = true
         return tagRepository.save(tag)
     }
 
@@ -40,11 +42,21 @@ class TagService {
 
     fun delete(id: Long) {
         val tag = findById(id)
-        return tagRepository.delete(tag)
+
+        if (postRepository.countByTags_Id(tag.id!!) == 0L) {
+            tagRepository.delete(tag)
+        } else {
+            tag.enabled = false
+            tagRepository.save(tag)
+        }
     }
 
-    fun findFilterPageable(page: Int, size: Int, search: String?): Page<Tag> {
-        return tagCriteria.findFilterPageable(page, size, search)
+    fun findPublicFilterPageable(page: Int, size: Int, search: String?): Page<Tag> {
+        return tagCriteria.findPublicFilterPageable(page, size, search)
+    }
+
+    fun findFilterPageable(page: Int, size: Int, search: String?, enabled: Boolean?): Page<Tag> {
+        return tagCriteria.findFilterPageable(page, size, search, enabled)
     }
 
 }
